Remove double shall in Section 6.1.6
Section 6.1.6 says that "The CA SHALL confirm that the value of the public exponent e shall be an odd positive integer," which suggests that the CA has to confirm that the requirement exists rather than confirming that the exponent meets the requirement. The CA needs to confirm that e "is" an odd positive integer, not that it "shall be" one.